Make SourceFileWithReferences `abstract`, so we can define `getReferences` once
