+++ rewrite-core/src/main/java/org/openrewrite/text/PlainText.java
@@ -16,12 +16,10 @@
 package org.openrewrite.text;
 
 import lombok.*;
-import lombok.experimental.NonFinal;
 import org.jspecify.annotations.Nullable;
 import org.openrewrite.*;
 import org.openrewrite.marker.Markers;
 
-import java.lang.ref.SoftReference;
 import java.nio.charset.Charset;
 import java.nio.charset.StandardCharsets;
 import java.nio.file.Path;
@@ -38,7 +36,8 @@ import static org.openrewrite.internal.ListUtils.nullIfEmpty;
 @Value
 @Builder
 @AllArgsConstructor
-public class PlainText implements SourceFileWithReferences, Tree {
+@EqualsAndHashCode(callSuper = false)
+public class PlainText extends SourceFileWithReferences implements Tree {
 
     @Builder.Default
     @With
@@ -84,7 +83,7 @@ public class PlainText implements SourceFileWithReferences, Tree {
     public PlainText withText(String text) {
         if (!text.equals(this.text)) {
             return new PlainText(this.id, this.sourcePath, this.markers, this.charsetName, this.charsetBomMarked,
-                    this.fileAttributes, this.checksum, text, this.snippets, this.references);
+                    this.fileAttributes, this.checksum, text, this.snippets);
         }
         return this;
     }
@@ -125,28 +124,7 @@ public class PlainText implements SourceFileWithReferences, Tree {
         if (this.snippets == snippets) {
             return this;
         }
-        return new PlainText(id, sourcePath, markers, charsetName, charsetBomMarked, fileAttributes, checksum, text, snippets, references);
-    }
-
-    @Nullable
-    @NonFinal
-    @ToString.Exclude
-    transient SoftReference<References> references;
-
-    @Override
-    public References getReferences() {
-        References cache;
-        if (this.references == null) {
-            cache = References.build(this);
-            this.references = new SoftReference<>(cache);
-        } else {
-            cache = this.references.get();
-            if (cache == null || cache.getSourceFile() != this) {
-                cache = References.build(this);
-                this.references = new SoftReference<>(cache);
-            }
-        }
-        return cache;
+        return new PlainText(id, sourcePath, markers, charsetName, charsetBomMarked, fileAttributes, checksum, text, snippets);
     }
 
     @Value
